+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.16
 -----------------
 - Add ETag cache system in Warehouse.php & Modules.php
+- CSS responsive image in stylesheet.css
 
 Changes in 2.9.15
 -----------------